Fix Google Spreadsheet Get
Reverted key change to previous commit. This turned out to not be the issue that broke the script.  Instead, the problem was caused because cors.io went down.  By passing in `?alt=json-in-script` google will automatically return the contents as jsonp so we don't need to run through a proxy.